apps/api/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
celery[redis]==5.3.6
python-dotenv==1.0.1
pydantic-settings==2.2.1
pgvector==0.2.5
httpx==0.27.0
greenlet==3.0.3
requests==2.31.0
beautifulsoup4==4.12.3
pdfplumber==0.10.3
pandas==2.2.0
sentence-transformers==2.5.1
redis==5.0.1
nltk==3.8.1
scikit-learn==1.4.1.post1
python-pptx==0.6.23
matplotlib==3.8.3
pytest==8.0.2
httpx==0.27.0
pytest-asyncio==0.23.5
